requirements.txt
# It's recommended to generate this file with 'pip freeze > requirements.txt'
# to get the exact versions from your environment.

fastapi
uvicorn
gunicorn # For production deployment
PyMuPDF
python-dotenv
google-cloud-translate
reportlab
firebase-admin
google-generativeai
networkx
firebase-admin
python-dotenv
